Issue #30 Implement simple Oauth authentication/authorization
This implements a simple system for authenticating users through Oauth.
Currently it only supports a stups setup, but later it should be
possible to use any oauth service for authentication.

Authorization is handled by defining a list of admin users in
application.yml or the equvalent environtment variable:

    pazuzu-registry:
      admins: user1,user2

The endpoint `/api/**` for methods `PUT` and `DELETE` are only open for
admins, and `POST` is only allowed by authenticated users.
`GET` is possible without login for any endpoint.

Closes #30